Add files via upload
capsFile.txt: a .txt file (really a .csv) containing pairs of US states and capitals in comma separated format.
capitalsRegex.py: a short python script that I used to extract the capitals from a python dictionary into the capsFile.txt file from a previous version of the code. It's not super useful for anything now, but it could technically be used to extract values from any hard-coded python dictionary into a CSV file. 
randomQuizGenerator.py: this is the main program file. This file handles both instances of the program. The user can use this file to either generate however many quizzes they want, or to take a practice quiz in the console.